chore(PageBrowser): move bread crumbs into their own component
