+++ src/components/Page/PageBrowser.vue
@@ -9,51 +9,12 @@
 		size="normal"
 		class="page-browser"
 		@closing="onClose">
-		<span class="crumbs">
-			<div v-if="!selectedCollective || !selectedCollective.isPageShare" class="crumbs-home">
-				<NcButton
-					variant="tertiary"
-					:aria-label="t('collectives', 'Breadcrumb for list of collectives')"
-					:disabled="!selectedCollective"
-					class="crumb-button home"
-					@click="onClickCollectivesList">
-					<template #icon>
-						<CollectivesIcon :size="20" />
-					</template>
-					{{ collectivesCrumbString }}
-				</NcButton>
-				<ChevronRightIcon :size="20" />
-			</div>
-			<template v-if="selectedCollective">
-				<div class="crumbs-level">
-					<NcButton
-						variant="tertiary"
-						:aria-label="collectiveBreadcrumbAriaLabel"
-						:disabled="pageCrumbs.length === 0"
-						class="crumb-button"
-						@click="onClickCollectiveHome">
-						<template v-if="collectiveBreadcrumbEmoji" #icon>
-							{{ collectiveBreadcrumbEmoji }}
-						</template>
-						{{ collectiveBreadcrumbTitle }}
-					</NcButton>
-				</div>
-				<div
-					v-for="(page, index) in pageCrumbs"
-					:key="page.id"
-					class="crumbs-level">
-					<ChevronRightIcon :size="20" />
-					<NcButton
-						variant="tertiary"
-						:aria-label="t('collectives', 'Breadcrumb, navigate to page {page}', { page: page.title })"
-						:disabled="(index + 1) === pageCrumbs.length"
-						class="crumb-button"
-						@click="onClickPage(page)">
-						{{ page.title }}
-					</NcButton>
-				</div>
-			</template>
-		</span>
+		<BreadCrumbs
+			:selectedCollective
+			:pageCrumbs
+			@clickCollectivesList="onClickCollectivesList"
+			@clickCollectiveHome="onClickCollectiveHome"
+			@onClickPage="onClickPage" />
 		<div class="browser-list">
 			<ul v-if="!selectedCollective">
 				<li
@@ -153,10 +114,10 @@ import NcDialog from '@nextcloud/vue/components/NcDialog'
 import NcLoadingIcon from '@nextcloud/vue/components/NcLoadingIcon'
 import ArrowDownIcon from 'vue-material-design-icons/ArrowDown.vue'
 import ArrowUpIcon from 'vue-material-design-icons/ArrowUp.vue'
-import ChevronRightIcon from 'vue-material-design-icons/ChevronRight.vue'
 import CollectivesIcon from '../Icon/CollectivesIcon.vue'
 import PageIcon from '../Icon/PageIcon.vue'
 import SkeletonLoading from '../SkeletonLoading.vue'
+import BreadCrumbs from './PageBrowser/BreadCrumbs.vue'
 import { useCollectivesStore } from '../../stores/collectives.js'
 import { usePagesStore } from '../../stores/pages.js'
 import { useRootStore } from '../../stores/root.js'
@@ -167,11 +128,11 @@ export default {
 	components: {
 		ArrowDownIcon,
 		ArrowUpIcon,
-		ChevronRightIcon,
 		CollectivesIcon,
 		NcButton,
 		NcDialog,
 		NcLoadingIcon,
+		BreadCrumbs,
 		PageIcon,
 		SkeletonLoading,
 	},
@@ -268,30 +229,6 @@ export default {
 			return this.isCurrentCollective
 				? this.pageParents(this.selectedPageId)
 				: this.pageParentsForCollective(this.selectedCollective, this.selectedPageId)
-		},
-
-		collectivesCrumbString() {
-			return this.selectedCollective
-				? ''
-				: t('collectives', 'All collectives')
-		},
-
-		collectiveBreadcrumbAriaLabel() {
-			return this.selectedCollective.isPageShare
-				? t('collectives', 'Breadcrumb for page {name}', { name: this.rootPage.title })
-				: t('collectives', 'Breadcrumb for collective {name}', { name: this.selectedCollective.name })
-		},
-
-		collectiveBreadcrumbEmoji() {
-			return this.selectedCollective.isPageShare
-				? this.rootPage.emoji
-				: this.selectedCollective.emoji
-		},
-
-		collectiveBreadcrumbTitle() {
-			return this.selectedCollective.isPageShare
-				? this.rootPage.title
-				: this.selectedCollective.name
 		},
 
 		movePageString() {
@@ -449,52 +386,6 @@ export default {
 .page-browser {
 	display: flex;
 	flex-direction: column;
-}
-
-.crumbs {
-	color: var(--color-text-maxcontrast);
-	display: inline-flex;
-	padding-right: 0;
-	padding-bottom: 8px;
-
-	div {
-		display: flex;
-		text-overflow: ellipsis;
-		white-space: nowrap;
-		overflow: hidden;
-		max-width: 300px;
-
-		.crumb-button {
-			color: var(--color-text-maxcontrast);
-
-			&.home {
-				padding-left: 0;
-				// Remove padding, add margin to not make the button bigger
-				padding-right: 0;
-				margin-right: var(--button-padding);
-				font-weight: bold;
-			}
-		}
-
-		&.crumbs-home {
-			flex-shrink: 0;
-		}
-
-		&.crumbs-level {
-			display: inline-flex;
-			min-width: 65px;
-
-			&:last-child {
-				flex-shrink: 0;
-			}
-		}
-
-		&:last-child {
-			.crumb-button {
-				color: var(--color-main-text);
-			}
-		}
-	}
 }
 
 .browser-list {
